move CharacterTypes enum to own file
