+++ backend/src/test/java/it/chalmers/gamma/api/OauthTests.java
@@ -1,6 +1,6 @@
 package it.chalmers.gamma.api;
 
-import static it.chalmers.gamma.utils.GenerationUtils.CharacterTypes.LOWERCASE;
+import static it.chalmers.gamma.utils.CharacterTypes.LOWERCASE;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
