+++ backend/src/test/java/it/chalmers/gamma/factories/MockITClientFactory.java
@@ -2,6 +2,7 @@ package it.chalmers.gamma.factories;
 
 import it.chalmers.gamma.domain.dto.access.ITClientDTO;
 import it.chalmers.gamma.service.ITClientService;
+import it.chalmers.gamma.utils.CharacterTypes;
 import it.chalmers.gamma.utils.GenerationUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -15,7 +16,7 @@ public class MockITClientFactory {
     public ITClientDTO generateClient(String redirect) {
         return new ITClientDTO(
             redirect,
-                GenerationUtils.generateRandomString(20, GenerationUtils.CharacterTypes.LOWERCASE),
+                GenerationUtils.generateRandomString(20, CharacterTypes.LOWERCASE),
                 GenerationUtils.generateText()
         );
     }
