+++ backend/src/test/java/it/chalmers/gamma/factories/MockWebsiteFactory.java
@@ -2,6 +2,7 @@ package it.chalmers.gamma.factories;
 
 import it.chalmers.gamma.domain.dto.website.WebsiteDTO;
 import it.chalmers.gamma.service.WebsiteService;
+import it.chalmers.gamma.utils.CharacterTypes;
 import it.chalmers.gamma.utils.GenerationUtils;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,8 +17,8 @@ public class MockWebsiteFactory {
     public WebsiteDTO generateWebsite() {
         return new WebsiteDTO(
                 UUID.randomUUID(),
-                GenerationUtils.generateRandomString(20, GenerationUtils.CharacterTypes.LOWERCASE),
-                GenerationUtils.generateRandomString(20, GenerationUtils.CharacterTypes.LOWERCASE)
+                GenerationUtils.generateRandomString(20, CharacterTypes.LOWERCASE),
+                GenerationUtils.generateRandomString(20, CharacterTypes.LOWERCASE)
         );
     }
