+++ backend/src/test/java/it/chalmers/gamma/utils/GenerationUtils.java
@@ -11,28 +11,6 @@ public final class GenerationUtils {
 
     }
 
-    public enum CharacterTypes {
-        UPPERCASE("ABCDEFGHIJKLMNOPQRSTUVWXYZ"),
-        LOWERCASE("abcdefghijklmnopqrstuvwxyz"),
-        NUMBERS("123456789"),
-        SPECIALS("!@#$%&()+=[]|/?><");
-
-        private String characters;
-
-        CharacterTypes(String characters) {
-            this.characters = characters;
-        }
-
-        public String getCharacters() {
-            return this.characters;
-        }
-
-        public static CharacterTypes[] allValues() {
-            return new CharacterTypes[]{
-                    UPPERCASE, LOWERCASE, NUMBERS, SPECIALS
-            };
-        }
-    }
 
     /**
      * Generate String of custom length.
